Partition sampling in Bayesian RaCE-NMA models (internal use only)

This function implements a reversible jump MCMC procedure for updating the parameter partition in Bayesian Rank-Clustered Estimation for Network Meta-Analysis models in the case when we assume correlation among treatment effects. For internal use only.

sample_partition_correlation(
  ybar,
  J,
  nu,
  g,
  K,
  mu0,
  sigma0,
  cov,
  tau = tau,
  b_g = 0.5,
  d_g = 0.5
)

Arguments

ybar

A vector of estimated average relative intervention effects based on a previous NMA. The jth entry is the effect of intervention j.

J

A numeric indicating the total number of interventions being compared.

nu

A vector indicating current values for nu in the Gibbs sampler.

g

A vector indicating current values for g in the Gibbs sampler.

K

A vector indicating current values for K in the Gibbs sampler.

mu0

The hyperparameter mu0, usually specified as the grand mean of the average intervention effects.

sigma0

The hyperparameter sigma_0, usually a large number as to be minimally informative.

cov

A variance covariance matrix of relative intervention effects based on a previous NMA. The (i,j) entry is the covariane between intervention i and j's effects.

tau

The standard deviation of the Metropolis Hastings proposal distribution.

b_g

The probability of "birth"ing a new partition cluster, if possible. Default is 0.5.

d_g

The probability of "death"ing an existing partition cluster, if possible. Default is 0.5.

Value

A list containing updated values for g, nu, and K.